requirements.txt
flask
python-dotenv
markdown
debugpy
